Moved AUTH command from the acl module to the connection module. Added echo command to the list of commands that are skipped on ACL authorization.
